refactor: re-arrange e2e directory
