+++ test/e2e/helper/util/clipboard_util.ts
